+++ src/tools.cpp
@@ -28,7 +28,7 @@ VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
   }
 
   //accumulate squared residuals
-  for (int i = 0; i < estimations.size(); ++i) {
+  for (size_t i = 0; i < estimations.size(); ++i) {
     VectorXd residual = estimations[i] - ground_truth[i];
     rmse = rmse.array() + residual.array() * residual.array();
   }
